+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## [1.6.0]
 ### Added
  - All HIV genes
+ - Add `-g` to julietflow to clip to a certain region
 
 ### Changed
  - Julietflow keep tmp dir renamed to `-z`, as `-t` is for target sequence
  - New HTML look and feel
  - Reverse overcorrection fix, not prime time ready yet
+ - Fix offset when config is missing
 
 ## [1.5.0]
 ### Added