+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
  - All HIV genes
  - Add `-g` to julietflow to clip to a certain region
+ - Add mixdata script
 
 ### Changed
  - Julietflow keep tmp dir renamed to `-z`, as `-t` is for target sequence